Removed note naming requirement
